safari-600.7-branch/LayoutTests/http/tests/notifications/legacy/request.html

<!DOCTYPE html>
<script src="/resources/js-test-pre.js"></script>
<script src="/resources/notifications-test-pre.js"></script>
<p id="description"></p>
<div id="console"></div>
<script>
description("This test checks that a request for permission is made.");

function step1() {
    debug("Requesting permission with default permission");
    shouldBe("window.webkitNotifications.checkPermission()", "1");
    webkitNotifications.requestPermission(function() {
        shouldBe("window.webkitNotifications.checkPermission()", "0");
        setTimeout(step2, 0);
    });
}

function step2() {
    testRunner.denyWebNotificationPermission(testURL);
    debug("Requesting permission with non-default permission");
    webkitNotifications.requestPermission(function() {
        shouldBe("window.webkitNotifications.checkPermission()", "2");
        testCompleted();
    });
}

if (window.testRunner) {
    testRunner.waitUntilDone();
    step1();
}

setTimeout(testCompleted, 100);
</script>
