safari-600.7-branch/LayoutTests/http/tests/notifications/legacy/request-expected.txt

This test checks that a request for permission is made.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Requesting permission with default permission
PASS window.webkitNotifications.checkPermission() is 1
PASS window.webkitNotifications.checkPermission() is 0
Requesting permission with non-default permission
PASS window.webkitNotifications.checkPermission() is 2
PASS successfullyParsed is true

TEST COMPLETE
